feat: `multiexp_serial` skips doubling when all bits are zero
For each `c` bit "segment", we get the bits from each coefficient in
`coeffs` ahead of time and determine what the largest segment is. We use
this to determine the largest nonzero segment and skip higher segments.

Before this, we always looked at all 256 bits and did a doubling
regardless. Now, if all the coefficients in the MSM are 64 bits, we will
detect this first and only do pippenger up to 64 bits. This only saves
doublings because the pippenger buckets for addition were already
dynamically created.